Removing the hugo-material theme as it is no longer needed.
This directory included image files which after investigation
were deemed not in current use on the website and could be safely
removed. If this assertion is wrong and they are needed later,
this commit can be reverted without harm to the current theme,
or the images can be retrieved from git history.