Add dark mode with toggler switch in nav.
Default mode is either system preference if supported, or light mode if
not.

Highlight.js removed from config.toml as it did not seem to be in use.
The "monokai" theme for codeblocks has been changed to "pygments" theme
to support dark/light mode switching. It's not amazing on dark mode but
it's better than eye strain.